+++ project1/src/sortStr.h
@@ -4,9 +4,9 @@
 #include <iostream>
 #include <string.h>
 
-void merge_sort(long n, std::string path = "./input_string.txt");
-void quick_sort(long n, std::string path = "./input_string.txt");
-void heap_sort(long n, std::string path = "./input_string.txt");
-void insertion_sort(long n, std::string path = "./input_string.txt");
+void merge_sort(int n, std::string path = "./input_string.txt");
+void quick_sort(int n, std::string path = "./input_string.txt");
+void heap_sort(int n, std::string path = "./input_string.txt");
+void insertion_sort(int n, std::string path = "./input_string.txt");
 
 #endif
